requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
aniso8601==9.0.1
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.3.2
Automat==20.2.0
backports.zoneinfo==0.2.1
binance-connector==1.12.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cheroot==8.6.0
click==8.1.3
colorama==0.4.4
constantly==15.1.0
cryptography==37.0.1
cycler==0.11.0
dateparser==1.1.1
Deprecated==1.2.13
Flask==2.1.2
Flask-RESTful==0.3.9
fonttools==4.33.3
frozenlist==1.3.0
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.12.0
incremental==21.3.0
itsdangerous==2.1.2
jaraco.functools==3.5.0
Jinja2==3.1.2
kiwisolver==1.4.2
MarkupSafe==2.1.1
matplotlib==3.5.1
more-itertools==8.12.0
multidict==6.0.2
networkx==2.8
numpy==1.22.3
packaging==21.3
Pillow==9.1.0
psutil==5.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycoingecko==2.2.0
pycparser==2.21
pyOpenSSL==22.0.0
pyparsing==3.0.8
python-binance==1.0.16
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
redis==4.2.2
regex==2022.3.2
requests==2.27.1
service-identity==21.1.0
six==1.16.0
Twisted==22.4.0
txaio==22.2.1
typing-extensions==4.2.0
tzdata==2022.1
tzlocal==4.2
ujson==5.2.0
unicorn-binance-websocket-api==1.41.0
unicorn-fy==0.12.2
urllib3==1.26.9
websocket-client==1.3.2
websockets==10.3
Werkzeug==2.1.2
wrapt==1.14.1
yarl==1.7.2
zipp==3.8.1
zope.interface==5.4.0
